refactor: Move <SelectCategoryInput/> to @chainlit/components
