+++ libs/components/src/inputs/index.tsx
@@ -1,7 +1,8 @@
 export { InputLabel } from './InputLabel';
 export { InputStateHandler } from './InputStateHandler';
-export { TextInput } from './TextInput';
+export { SelectCategoryInput } from './selects/SelectCategoryInput';
 export { SelectInput } from './selects/SelectInput';
+export { TextInput } from './TextInput';
 
 export type { SelectInputProps, SelectItem } from './selects/SelectInput';
 export type { TextInputProps } from './TextInput';
